refactor(api): extract rate limiter storage backends into their own module
Move RateLimitBackend, InMemoryBackend, and RedisBackend out of the 1288-line
rate_limit.py into rate_limit_backends.py (302 lines), re-exported so existing
imports keep working. rate_limit.py drops to ~1014 lines. Covered by the new
rate-limiter unit tests and the full backend suite (both green). The
RateLimiter/GenerationRateLimiter dedup is deferred until those paths have
dedicated test coverage.
